Revert "Making this a Kotlin library rather than an Android library."
